requirements.txt
numpy==2.3.0
opencv-python==4.11.0.86
pyserial==3.5
